Refactor: Enum 패키지 분리 #18
