+++ src/main/java/com/cojac/storyteller/domain/setting/FontSize.java
@@ -1,4 +1,4 @@
-package com.cojac.storyteller.domain;
+package com.cojac.storyteller.domain.setting;
 
 public enum FontSize {
     SMALL("작게"), MEDIUM("보통"), LARGE("크게");
@@ -9,8 +9,5 @@ public enum FontSize {
         this.description = description;
     }
 
-    public String getDescription() {
-        return description;
-    }
 }
